requirements.txt
exa-py>=1.7.1
python-dotenv>=1.0.0
streamlit>=1.29.0
pandas>=2.1.0
altair>=5.2.0
